Moved conversion functionality to convert-schema.js
